[Linux] Eager assembly load of managed profiler code (#469)
The main mechanism for loading the managed profiler code was introduced in #462 (commit 4ce932f).

This PR turns on the feature for Linux. The main components are embedding the resources into the shared library, retrieving the embedded resources in the native profiler, and adding the correct PInvoke from the injected assembly load.