Changes to fix an issue with recent MSI changes. (#522)
In recent changes, we realized we needed to re-install the net45 files in the GAC to allow IIS automatic instrumentation to succeed so we made the change. However, as a result, the net45 files were not placed on-disk in a subdirectory of `DD_DOTNET_TRACER_HOME`. This change fixes that by installing the files twice: once into the GAC and once on-disk.