Fix log4net.Util.PropertiesDictionary SerializationException with DD_LOGS_INJECTION=true (#551)
Fixes:
SerializationException: Type is not resolved for member 'log4net.Util.PropertiesDictionary,log4net, Version=2.0.8.0, Culture=neutral, PublicKeyToken=669e0ddf0bb1aa2a'

Scenario:
This crash occurs when 3 conditions are met:

The user is running an ASP.NET application in IIS
The user references the Datadog.Trace.ClrProfiler.Managed NuGet package (which is required for the AspNet integration) up to and including 1.9.0
DD_LOGS_INJECTION=true
This occurs because at the time of the Asp.Net HttpModule startup, we initialize the Tracer and set default values of "0" for "dd.trace_id" and "dd.span_id". When the initialization is finished and control is passed back to IIS, an AppDomain transition occurs. At this time, the log4net.Util.PropertiesDictionary that held the dd.trace_id and dd.span_id properties must be passed across AppDomains because it was stored inside the System.Runtime.Remoting.Messaging.CallContext structure. When IIS receives the object, it must be deserialized, at which point it crashes because the log4net assembly cannot be found: it was app-local.

The solution is to not assign default values when the Tracer is created, only assigning values to dd.trace_id and dd.span_id in response to the changing state of active spans in the application.

Changes proposed in this pull request:

When DD_LOGS_INJECTION=true, do not set default values for dd.trace_id and dd.span_id when the Tracer is created, because it may not be safe to do so at that time.
Add unit tests to ensure this behavior
Add smoke test to replicate the IIS scenario of running another application in a separate AppDomain, with the log4net dependency, with DD_LOGS_INJECTION=true