Prevent AssemblyResolve Event from causing infinite recursion on mscorlib resource lookup (#573)
Modifies the AssemblyResolve event handler in the Datadog.Trace.ClrProfiler.Managed.Loader assembly to eagerly exit if the requested assembly is `mscorlib.resources` or `System.Private.CoreLib.resources` on .NET Framework and .NET Core, respectively.

This fixes a known issue where on a non-US locale, a lookup of an `mscorlib` resource invokes the AssemblyResolve event and our event handler causes infinite recursion and the .NET application to crash. This occurred because, in processing the event, the event handler would make a method call to `File.Exists`, which triggers yet another AssemblyResolve event for `mscorlib` and the cycle continues. Even though this has only been demonstrated to occur on .NET Framework and should not occur on .NET Core 2.0+ (see https://github.com/dotnet/coreclr/issues/12668), proactively guard against AssemblyResolve events for `System.Private.CoreLib.resources` as well.