AspNet HttpModule: Fix bug when adding to web.config (#584)
The previous implementation of the HttpModule did not behave correctly when it was invoked more than once in one ASP.NET request. Multiple scopes would be created and only the innermost one would be closed, so the root span would not be closed and the Datadog UI would show no traces. The new implementation registers the first HttpModule to hit the event handlers as the only one allowed to create/modify Scope objects, so only one scope is created and closed.

This fix is important to support customers who have other HttpModules in their system web.config that subscribe to the BeginRequest and would like our Tracer to start spans before hitting them. As a result, the customer will enter our HttpModule before the other HttpModule's in their web.config which causes our the ASP.NET application to create and invoke two instances of our HttpModule: once from the web.config declaration and once from the dynamic module after referencing our NuGet package.