Issue: When using the AspNetScopeManager, the closing of a span generated by automatic instrumentation doesn't get propagated correctly to the original ASP.NET synchronization context, causing later spans to have the wrong parent.
This occurs because we use `ConfigureAwait(false)` on all of our async continuations. When we create the new span we register it to the `HttpContext.Current` of the ASP.NET synchronization context, but when we close the span during the continuation we have lost the original ASP.NET synchronization context so the `HttpContext.Current` remains unchanged. Then, later spans generated in the ASP.NET synchronization context treat that span as being active and will always be a child of that one.

This commit adds a DadJoke endpoint to illustrate the above issue.