[ClrProfiler] Fix DbCommand.xxxAsync failures on 32-bit .NET Framework (#702)
See https://github.com/DataDog/dd-trace-dotnet/issues/662

## Issue
With the way our CLR Profiler and `Datadog.Trace.ClrProfiler.Managed.dll` is set up, the integration methods that we use for automatic instrumentation must only have method arguments whose types are .NET primitive types (e.g. `int`, `long`, `byte`) or `object`. To handle `xxxAsync` methods in .NET Framework libraries and third-party libraries that take a `System.Threading.CancellationToken` argument (which is a [value type](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/builtin-types/value-types), not a [reference type](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/reference-types)), the instrumentation method would specify the argument type as `object`, cast it to a `System.Threading.CancellationTokenSource`, and then use the `Token` property to access the original CancellationToken. While this has silently worked for our integrations on almost all platforms, it crashes on 32-bit .NET Framework when instrumenting the `System.Data.dll` assembly.

The failure likely occurs because when the instrumentation method is called, it expects the CancellationToken argument to be populated with a reference type of type `object` that has a reference to the **real** `System.Threading.CancellationToken` value. However, since `System.Threading.CancellationToken` is a value type, the value that is passed to the method is not a reference but the data itself. When the data is interpreted to be a reference and the variable is accessed, this can cause invalid memory accesses. The fix is to always [box](https://docs.microsoft.com/en-us/dotnet/csharp/programming-guide/types/boxing-and-unboxing) the `System.Threading.CancellationToken` struct so that an `object` is placed onto the stack for the method call and the instrumentation method can read it correctly.

## Changes
- Fix the CLR Profiler to add an additional `box <typeToken>` IL instruction when performing a method replacement and the last argument of the target method has the type `System.Threading.CancellationToken`
- Modify all integration methods to directly cast their cancellationToken argument to `System.Threading.CancellationToken` instead of manipulating it as a `System.Threading.CancellationTokenSource`
- Add Windows `integration-test` run for x86 (previously was only x64), which exercises this code path in the `Samples.SqlServer` application
- Add minor logging to change to log the bitness of the Profiler when it initializes
- Add further documentation in the IL Modification part of the method replacement code