[ClrProfiler] Fix missing instrumentation when DD_TRACE_DOMAIN_NEUTRAL_INSTRUMENTATION is enabled (#738)
### Issue
The environment variable `DD_TRACE_DOMAIN_NEUTRAL_INSTRUMENTATION` has been issued as a workaround when an application running in IIS and Azure App Services is missing .NET framework spans because the framework is loaded in a shared context. This override is only safe when the user can guarantee that only one application is being loaded in the Application Pool.

The issue today is that even when the environment variable `DD_TRACE_DOMAIN_NEUTRAL_INSTRUMENTATION` is set, we still do not perform instrumentation when our instrumentation assembly `Datadog.Trace.ClrProfiler.Managed` is loaded app-local. Since the user knows that there will only be one application in the Application Pool (meaning only one AppDomain will be instrumented), this is safe to do and we should enable that functionality.

### Changes proposed in this pull request:
- Add a code path in the profiler to instrument domain-neutral assemblies when the environment variable `DD_TRACE_DOMAIN_NEUTRAL_INSTRUMENTATION` is set and the `Datadog.Trace.ClrProfiler.Managed` assembly is loaded app-local
- Add a new .NET Framework sample to test this scenario that creates outbound HTTP requests using the framework-provided `System` and `System.Net.Http`
- Add integration test to run this sample and to run the sample under a bad configuration to make sure we don't crash with the new functionality