[ClrProfiler] Allow automatic instrumentation of domain-neutral assemblies to work safely (#750)
# Summary
The automatic instrumentation has had a long-standing issue of safely instrumenting domain-neutral assemblies. This occurs because in a .NET application with multiple AppDomains and multiple assembly loading policies (e.g. IIS), `Datadog.Trace.ClrProfiler.Managed` may be loaded domain-neutral in one AppDomain but not in another. To fix the immediate code-sharing issue, we had set the default behavior of the native profiler to not instrument methods originating from a domain-neutral assembly. However, this caused missing traces for integrations involving `AdoNet`, `HttpClient`, `WebRequest` and `Wcf` when they were loaded domain-neutral.

This PR provides a general fix for the situation that applies to .NET Framework 4.5.2+, which provides the `ICorProfilerCallback6::GetAssemblyReferences` API. This API allows a CLR profiler to inform the runtime to extend the assembly closure of an assembly that is about to be loaded, and we use this mechanism to add a new assembly reference to `Datadog.Trace.ClrProfiler.Managed` for assemblies we are interested in instrumenting. As a result, `Datadog.Trace.ClrProfiler.Managed` is taken into consideration when deciding whether to load the instrumented assembly domain-neutral or not. If `Datadog.Trace.ClrProfiler.Managed` cannot be loaded domain-neutral in that AppDomain, then the instrumented assembly is not loaded domain-neutral either. This solves the problem scenario altogether.


**Emphasis:** This only works on machines running .NET Framework 4.5.2+. For machines running a lower version of the .NET Framework, `DD_TRACE_DOMAIN_NEUTRAL_INSTRUMENTATION` may still be used to instrument domain-neutral assemblies but it still may only be used if there is only a single AppDomain in that .NET process.

This is a continuation of #711.

# Changes made in this pull request

## Tracer
- Remove `System.Net.Http` assembly reference from `Datadog.Trace`
  - In `Datadog.Trace.Agent.Api`, replace `HttpClient` implementation with `WebRequest`. Additionally, use an abstraction to represent the requests and responses from the Datadog agent
- Remove `System.Net.Http` assembly reference from `Datadog.Trace.ClrProfiler.Managed`
  - Use reflection to inject/extract `HttpHeaders` for distributed tracing
  - Modify `AspNetWebApi2Integration` and `HttpMessageHandlerIntegration` integration to remove static references to System.Net.Http types. Instead, use the `AsyncHelper.InvokeGenericTaskDelegate` helper to return Tasks with the correct generic instantiation that the caller expects
- Modify the `Datadog.Trace.ClrProfiler.Managed.Loader` AssemblyResolve event to never load `System.Net.Http`

## Profiler
- Add `ICorProfilerCallback6::GetAssemblyReferences` callback to allow the profiler to inform the runtime of the `Datadog.Trace.ClrProfiler.Managed` dependency we're going to add later, allowing us to safely instrument assemblies that would have otherwise been loaded domain-neutral

## CI
- Consolidate domain-neutral tests into `Samples.MultiDomainHost.Runner` with several `Samples.MultiDomainHost.App.*` sub-applications to test the `LoaderOptimization.MultiDomainHost` scenario that IIS runs, and add two more sub-applications to the test
- Modify test helpers to require full-name of test application (e.g. `Samples.MongoDB`)
- Modify test helpers to allow for custom TFM subfolder, to allow for multi-targeted domain-neutral test case