Fix "CallVirtual expected a StackExchange.Redis.RedisBase; found StackExchange.Redis.RedisBase" (#872)
### Issue
Today we still have an issue in the automatic instrumentation of `StackExchange.Redis.RedisBase.ExecuteAsync`. The first time this method is executed, the type of the object instance is saved and cached for speedup purposes. Unfortunately, this leads to errors when both `StackExchange.Redis` and `StackExchange.Redis.StrongName` libraries are brought into the app because they can both enter this integration method but the types are **not** identical because they are defined in separate assemblies.

### Changes
- Fix the `StackExchange.Redis.RedisBase.ExecuteAsync` method replacement to always perform a type lookup on the instance object
- Improve our test application to exercise async code paths when both `StackExchange.Redis` and `StackExchange.Redis.StrongName` libraries are used
- Extra: Small GraphQL edit which removes unused code path