Remove System.Diagnostics.DiagnosticSource from the .NET Framework build (#881)
After working on #880 (which focuses on System.Net.Http), it seems especially dangerous to even have an assembly reference on an assembly that can be provided by NuGet instead of the GAC. After further testing, the same System.Net.Http failure mode can occur with System.Diagnostics.DiagnosticSource if there is a binding redirect on the assembly and one of the assemblies in the application was compiled against the same version as the .NET Tracer (System.Diagnostics.DiagnosticSource Version=4.0.2.1).

Since we do not use System.Diagnostics.DiagnosticSource right now on .NET Framework, this change removes it from the build to get rid of the possibility of assembly loading issues. Properly loading this on .NET Framework will be addressed when we begin moving towards using the Activity class to propagate events.