Fix SecurityException when making cross-AppDomain method calls and using CallContext (#925)
### Error Message
`System.Security.SecurityException: Request for the permission of type 'System.Security.Permissions.SecurityPermission, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089' failed.`

### Issue
The issue manifests under three conditions:
1. The user is running .NET Tracer 1.19.2 on .NET Framework
2. The `net45` Datadog .NET Tracer is used (default scenario), which stores the ambient `Active` scope inside the CallContext because there is no AsyncLocal<T> type
3. A method call is made into a new AppDomain without the `SecurityPermissionFlag.RemotingConfiguration` enabled

### Background
For net45 where AsyncLocal<T> is not a built-in Framework class, we must propagate the ambient context using LogicalCallContext. One consequence of this is that we invoke the System.Runtime.Remoting infrastructure, and if a remote method call or cross-AppDomain method call is made, objects stored inside the LogicalCallContext are transmitted to the other side (aka the "server").

For distributed garbage collection, the "server" uses leases / sponsors to keep track of what objects are needed by "clients" (our original AppDomain) and should not be garbage collected. If we do not refresh the lease on objects stored inside the LogicalCallContext, the "server" will garbage collect them and disconnect them, meaning any accesses of the object inside the LogicalCallContext will throw a RemotingException.

The previous solution for this problem was to create a wrapper object called `DisposableObjectHandle` (which is the object stored in the CallContext and transmitted to the "server") and let it manage its Lifetime with the Remoting infrastructure, which includes calling into the `base.InitializeLifetimeService()` method which is `SecurityCritical`. However, if execution flows into an AppDomain that does not have the `SecurityPermissionFlag.RemotingConfiguration` enabled, this will result in a SecurityException.

### Fix
Modify the `DisposableObjectHandle` wrapper to return `null` for its `InitializeLifetimeService()` method, which will keep the object alive indefinitely and not require any method calls into `SecurityCritical` code. Then, only remove the object from the "server" when the ambient `Active` scope in the originating AppDomain changes. This method call will not be invoked inside the locked-down AppDomain but instead in the application AppDomain where `Datadog.Trace` is loaded from the GAC.