In Datadog.Trace.Tests.Sampling.RateLimiterTests.Limits_Approximately_To_Defaults, reduce the number of spans sent in each burst to unit-test rate limiter in a non-overloaded state. (#926)
In Datadog.Trace.Tests.Sampling.RateLimiterTests.Limits_Approximately_To_Defaults, reduce the number of spans sent in each burst to unit-test rate limiter in a non-overloaded state.

The current unit test sends approximately 800 spans a second whereas the default threshold is 100 spans a second. In this scenario, the rate limiter is more aggressive than anticipated and leads to transient test failures. Instead, reduce the load to 400 spans a second.