Fix InvalidProgramException when instrumenting a constrained virtual method call (#937)
### Error Message
`System.InvalidProgramException: Common Language Runtime detected an invalid program.`

### Issue
The issue occurs when the automatic instrumentation tries to redirect a constrained virtual method call. A constrained virtual method call is a `constrained` IL instruction followed by a `callvirt` IL instruction. This can occur when making a method call to an object whose type is a generic type variable. Since the automatic instrumentation modifies the original method call to a `call` IL instruction to invoke any one of the static integration methods, the `constrained` IL instruction is no longer valid because it does not prefix a `callvirt` IL instruction, and the result is an `InvalidProgramException`.

### Fix
When trying to apply automatic instrumentation, check if the original `callvirt` IL instruction is preceded by a `constrained` IL instruction and, if so, bail out on instrumentation in this specific method.

### Changes
- Implement profiler fix
- Add the constrained virtual method call scenario to the AdoNet sample apps, which simultaneously tests that we do not crash and will test that, in the future when CallTarget instrumentation is performed, we correctly instrument these calls
- Set the environment variable `DD_TRACE_NETSTANDARD_ENABLED=true` for all AdoNet sample apps by default
- Add to the process exclusion list in `devenv.bat`