[opentracing] Allow passing a Span Logger
One of the pitfalls of the OpenTracing implementation for the datadog
tracer is that it doesn't support logging within a span (span.Log()).

I totally understand that since "logs within a span" is not a feature
of the Datadog API platform, sending span logs inside a span object
would take quite a lot, and may not even be on the roadmap.

However, giving end users the ability to pass a "span logger" to the
OpenTracing tracer seemed like a good workaround: when instantiating
the Tracer, we can also provide a delegate that will be called when
span.Log() is.

Typically, this enables users to use their favorite logging library
to have the log appear on stdout, injecting Trace and Span ids in the
process. These logs can later be collected by the log agent, and they
will appear on the APM interface as one would expect :)

In a sense, it "fixes" (or at least offers a workaround for) this
issue: https://github.com/DataDog/dd-trace-dotnet/issues/909