Add automatic instrumentation for apps/libraries referencing System.Data 2.0.0.0 (#1000)
Fixes missing ADO.NET automatic instrumentation spans when running an application on .NET Framework 4.5 (or higher) and the application or a library was built against .NET Framework 2.0.

### Summary
When a .NET application or library is built for an older version of the .NET Framework it will have assembly references to that version of the framework assemblies (e.g. System.Data 2.0.0.0), but when it is run against a later version of the .NET Framework, the .NET Framework will automatically unify the framework assemblies to the current version. This means if a .NET Framework 2.0 application referenced System.Data 2.0.0.0 but it was run on .NET Framework 4.5, the runtime would actually load System.Data 4.0.0.0. In such a scenario, we should be performing automatic instrumentation because System.Data 4.0.0.0 is loaded. However, we currently do not instrument these call sites because our CallSite instrumentation can only see the version that the application/library referenced, which was System.Data 2.0.0.0, and our ADO.NET instrumentation has a set minimum AssemblyVersion of 4.0.0.0.

The proposed change is to lower the minimum AssemblyVersion of ADO.NET integration methods from 4.0.0.0 to 2.0.0.0 for the synchronous methods. This will be valid because:
1. The `ExecuteXXX` API's are identical between 2.0 and 4.0 (used for method replacement)
2. The `IDbCommand` interface is identical between 2.0 and 4.0 (used to populate span information)
3. If automatic instrumentation is on, then the user must be running .NET Framework 4.5 and, by extension, using System.Data 4.0.0.0 at runtime.

### Changes
- Lower the minimum AssemblyVersion for `DbCommand`, `IDbCommand` and `SqlCommand` integration methods from `4.0.0.0` to `2.0.0.0`
- Add a new Samples.SqlServer.NetFramework20 that uses a .NET Framework 2.0 library (compiled against System.Data 2.0.0.0)