Try building the solution instead of individual projects
It's often significantly faster to do so.
There are currently some projects which _were_ being built, which will no longer be built with this approach
Should we add them to the solution, or do they need to be excluded and explictly built?

- ExpenseItDemo.netcore.csproj
- NugetProfilerVersionMismatch.TypeLoadException.csproj
- Samples.WebRequest.csproj
- BenchmarkDotNet.Autogenerated.csproj
- Samples.AspNet472.LoaderOptimizationRegKey.csproj
- Samples.AspNetCore31.csproj

This commit also disables the build and restore stages when running dotnet test
As we know that's already been run