Copy managed profile to integration test output dir
If we exclude the profiler files in the sample apps, then this is where the samples try to load the profiler from. Without this, they will currently fail
Preferable to copying to sample output in CI as it uses a lot of drive space, and we're less concerned about locking.