Disable automatic logs injection during IIS PreStartInit phase (#1157)
### Issue
This change fixes a crash that occurs under the following conditions:
1. .NET Framework application is hosted in IIS
2. Automatic logs injection is enabled
3. Application uses the log4net logging library
4. Automatic instrumentation is triggered during the IIS PreStartInit phase

### Context
In order to store ambient information for the logging context on .NET Framework, log4net uses `CallContext`. The first time a property is stored, a log4net dictionary type is created and permanently stored in the `CallContext`, even if the dictionary is later found to be empty. This creates an error condition: If this is stored into the `CallContext` during the IIS PreStartInit phase (which occurs in the application-specific AppDomain), then a crash will occur when the phase is finished and execution resumes in the IIS default AppDomain.

### Solution
The solution is two-fold:

1. When automatic instrumentation is enabled, we already perform IL rewriting at the beginning of the IIS PreStartInit code path to load our managed code. Modify this code path to begin and end with a call to `AppDomain.SetData` to set a flag that indicates if the IIS PreStartInit phase is still running. This flag can be observed from elsewhere in the AppDomain and need not be on the same thread.
2. Register to a new `TraceStarted` event provided by the ScopeManager and register a callback to evaluate whether we're still running in this PreStartInit phase. It will either check for the AppDomain flag (if it was not-null) or it will fallback to checking the StackTrace (which is expensive). As soon as we detect false, we won't execute these checks again.

Optimizations to this strategy include never subscribing to the `TraceStarted` event if one of the following conditions is met at Tracer startup:
1. Automatic instrumentation is enabled and the IIS PreStartInit flag is already false
2. No automatic instrumentation is enabled and the process name is not `w3wp` or `iisexpress`