Implement log rate limiting as per RFC (#1153)
* Replace Log.Verbose with Log.Debug

* Update the minimum log level to Debug

The RFC only uses Debug, Warn, and Error. We add Information, for  emitting non-error logs (such as configuration status logs)

* Update some log levels to closer align with the RFC

* Add IDataDogLogger wrapper around ILogger

We'll use this to control which log levels etc are available, add rate limiting etc.
For now, implements all currently used APIs of ILogger

* Add method for retrieving IDatadogLogger

Unfortunately, despite being internal, we can't change the DatadogLogging class as it will cause MethodNotFound exceptions when there's a mismatch between Manual and Automatic packages.
Instead, mark the old methods obsolete, and add new preferred methods to use.

* Replace usages of ILogger with IDatadogLogger

* Make IDatadogLogger interface consistent across log levels

Ensuring we don't box and allocate unnecessarily makes the interface quite verbose unfortunately.
Replace use of params. This makes for an awkward interface for callers when many parameters are used, as you must pre-allocate the array at the call site.
Removing the use of params is necessary so that we can rate limit using [CallerFilePath], which requires optional parameters, and so are not compatible with params arguments

* Add [CallerLineNumber] and [CallerFilePath] attributes

Required for the log rate limiting RFC.
Unfortunately, means that some logging calls will be ambiguous unless you specify generic arguments, which is not ideal.
Placed the LineNumber attribute first, as more common to have a string argument than an int argument

* Add LogRateLimiter to ensure we only write a given log X times per second

Based on the APM Tracer Logging RFC
The default (null) implementation doesn't filter any log messages
All log messages that are written indicate how many previous instances were skipped

* Allow controlling rate limiting with DD_LOGGING_RATE, as per RFC

* Replace SafeLogError with Error

As we now have a wrapper for the ILogger, no reason to have a separate logging extension, as can make all logging events safe
The extension was not compatible with the CallerMemberName attribute so would need updating anyway
Confirmed with a microbenchmark that the extra overhead of adding the try-catch for every log event should be negligible + it's only there when the log level is enabled anyway

* Remove unused extension

Doesn't add anything over calling _logger.Error directly, and hides caller member name attributes

* Propogate caller member attributes in logging extensions

Necessary for rate limiter to throttle based on original call site, otherwise all logs that call the extension would be throtted together

* Handle circular static logger initialization in EnvironmentHelpers

DatadogLogging creates a shared logger instance. Creating this instance requires using `EnvironmentHelpers`.
This in turn tries to get the `DataDogLogging.SharedLogger`
This is initially set to a "null" logging implementation, and will remain as such for the lifetime of the program
Using a `Lazy<>` means the logger is only retrieved when ther's an exception in `EnvironmentHelpers`, hopefully after `SharedLogger` has been updated

* Rollback CallerMemberAttributes in CallTargetInvoker.LogException

Not entirely sure how to get these to hook to work, or if it's even possible...

* Replace empty array with Datadog.Trace.Util.ArrayHelper.Empty

Co-authored-by: Kevin Gosse <krix33@gmail.com>

* No need to use Select( x=> x.ToSring())

Co-authored-by: Kevin Gosse <krix33@gmail.com>

* PR review fixes

Revert log level change
Fix rate value guard
Don't add "skip count" message if skipCount == 0
Add comments
Fix usage of Clock in tests
LogRateBucketInfo can be readonly struct
Use struct as key for dictionary
Add escape hatch if CallerLineNumber attributes fails for some reason

* Fix merge issue

* Fix rate limit

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>

* Bail-out of LogRateLimiter if either filepath or line number is empty

Co-authored-by: Kevin Gosse <krix33@gmail.com>
Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>