Fix Datadog.Trace.ClrProfiler.IntegrationTests.StackExchangeRedisTests.SubmitsTraces
The KeyExpire API calls were expressed as DateTime's. This introduced variability because the DateTime for the final redis command could sometimes be represented as seconds and other times only as milliseconds, leading to occasional EXPIREAT (seconds) commands instead of the expected PEXPIREAT (milliseconds).

The fix is to make the API call with a specific TimeSpan instead so the resulting redis command is always PEXPIRE.