Fix ASP.NET core DiagnosticObserver recording wrong status code for unhandled exceptions
If there's no error handling middleware, the HttpServer (e.g. Kestrel) handles the exception, and updates the response status code. Although the unhandled exception callback in diagnostic observer is invoked, the status code still has a 200 status code at this point, as Kestrel doesn't update it to 500/4X error until later

To work around this, we assume a status code of 500 (or use the value in the internal BadHttpRequestException), and _don't_ update the status code on request end.