Add additional ASP.NET Core tests + fix response code bug (#1269)
* Move SerializableDictionary to TestHelpers library

* Add additional AspNetCoreDiagnosticObserver tests

Adds tests for tags set (none) and resource names, in preparation for updating these later behind a feature flag

Required so we can reference them directly from test apps

* Add testing of response and span status code in Diagnostic Source

* Add a Razor Pages sample application for diagnostic observer tests

* Add tests for AspNetCore RazorPages

* Add tests for ApiController endpoints

* Add testing of unhandled exception handling

Tests several different scenarios:
- No error handling middleware, HttpServer handles the exception
- Standard developer exception page
- Error handling middleware with pipeline reexecution
- Custom Handler, which only catches a subset of errors

Used multiple Mapped pipelines to make testing the different paths easier. Hardest thing was hitting the edge case path in Kestrel for when the request is malformed, as that was made internal in 3.0, but is still worth testing I think

* Add tests for exceptions in non-mvc pipelines

* Fix ASP.NET core DiagnosticObserver recording wrong status code for unhandled exceptions

If there's no error handling middleware, the HttpServer (e.g. Kestrel) handles the exception, and updates the response status code. Although the unhandled exception callback in diagnostic observer is invoked, the status code still has a 200 status code at this point, as Kestrel doesn't update it to 500/4X error until later

To work around this, we assume a status code of 500 (or use the value in the internal BadHttpRequestException), and _don't_ update the status code on request end.

* Skip re-execute tests for now

It's not possible to fix these cases without more extensive work. These can be fixed at the same time as we fix the overall resource names.

* Move tests to integration library

* Fix build configurations

* Strong name the RazorPages project that is directly referenced from test project

* Set the application parts for TestServer tests

Ensures we don't accidentally pick up the Razor Pages app

* Fix Datadog.Trace.IntegrationTests output type

* Update StatusCode API based on PR comments