I looked to port https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/pull/71
However, I don't think we necessarily need to take the changes from there, which can be summed up as adding a Metrics and Tags property on the ITags interface, to allow their Zipkin exporter to read the tags/metrics.

In the end, I just ported one unit test and modified it to work with our structure.