Add AssemblyLoadContext and Assembly.LoadFile testing to ServiceStackRedis sample (#1337)
To further test resiliency against assembly loading issues, add AssemblyLoadContext and Assembly.LoadFile testing to the ServiceStackRedis sample. This augments the existing ADO.NET tests that do a similar thing, except in the ServiceStack.Redis library, all of the types are defined in the third-party library, unlike ADO.NET where some types are defined in the framework-provided System.Data assembly.