Update Uri cleaning algorithm to be more aggressive (#1327)
* Update Uri cleaning algorithm to be more aggressive

Replaces any hex-ish (alpha or `-`) string with `?`.
If it's numbers and dashes only, always replaces
If it contains alpha, must be at least 16 chars long and contain a number
Added in back-compat way, that avoids changing internal (public) APIs

* Conditionally enable the new URI cleaning algorithm in ASP.NET

Enabled when the new route names flag is enabled
Added a note that we're currently using the _legacy_ algorithm still in the generic scope factory for _outbound_ requests. We may want to update this, but wasn't sure if we should do it under the same flag, as it's _kinda_ unrelated

* Simplify implementation

* Mark older UriHelpers methods Obsolete (kept for back compat)

Add documentation to CleanUri method
Update tests to use non-obsolete methods

* Use new Uri cleaning in AspNetCore resources when feature flag enabled

* Remove the feature-flag behaviour i.e. always use the new Id replacement algorithm

We decided to take this approach as it's simpler, and in practice the cleaning algorithm should only be used when there's a 404 anyway. If customers find the new algorithm is causing them problems, opting in to the new route-based resource names should fix it.

The one other place this cleaning algorithm is used is in the ScopeFactory for outbound HTTP requests. There's no workaround there, but the chances of issues are small, so we're ok with the tradeoff for simplicity and performance